Broke out fastqc functionality in separate script
