+++ qtp_target_gene/summary.py
@@ -15,6 +15,8 @@ from StringIO import StringIO
 
 
 from qiita_ware.demux import stats as demux_stats
+from qtp_target_gene.fastqc import fastqc
+
 import matplotlib.pyplot as plt
 
 
@@ -192,199 +194,3 @@ def _summary_demultiplexed(artifact_type, filepaths):
 
     return artifact_information
 
-
-def generate_fastqc_commands(forward_seqs, reverse_seqs, map_file, out_dir):
-    """Generates the FastQC commands
-
-    Parameters
-    ----------
-    seqs : list of str
-        The list of seqs filepaths
-    map_file : str
-        The path to the mapping file
-    out_dir : str
-        The job output directory
-
-    Returns
-    -------
-    list of str
-        The FastQC commands
-
-    Raises
-    ------
-
-    Notes
-    -----
-    This is presently reproducing functionality from the kneaddata pipeline
-    in qp_shotgun. 
-    """
-
-    samples = make_read_pairs_per_sample(forward_seqs, reverse_seqs, map_file)
-
-    cmds = []
-    prefixes = []
-
-    fps = []
-    for run_prefix, sample, f_fp, r_fp in samples:
-        prefixes.append(run_prefix)
-        if r_fp is None:
-            cmds.append('mkdir -p %s; fastqc --outdir "%s" %s' %
-                        (join(out_dir, run_prefix), join(out_dir, run_prefix),
-                         f_fp))
-            fps.append((f_fp, None))
-        else:
-            cmds.append('mkdir -p %s; fastqc --outdir "%s" %s %s' %
-                        (join(out_dir, run_prefix), join(out_dir, run_prefix),
-                         f_fp, r_fp))
-            fps.append((f_fp, r_fp))
-
-    return cmds, samples
-
-
-def generate_multiqc_commands(file_dir, out_dir):
-    """Generates the multiqc commands
-
-    Parameters
-    ----------
-    file_dir : str
-        Filepath to directory of QC info
-    out_dir : str
-        The job output directory
-
-    Returns
-    -------
-    list of str
-        The multiqc commands
-
-    Notes
-    -----
-    """
-
-    # make input file list
-
-    cmds = []
-
-    cmds.append('multiqc --outdir %s --filename MultiQC.html %s'
-                '%s' % (out_dir, file_dir))
-
-    return cmds
-
-
-def _guess_fastqc_filename(fp):
-    f_p = basename(fp)
-
-    exts = ['.fastq', '.fq', '.gz', '.gzip']
-    while splitext(f_p)[1] in exts:
-        f_p = splitext(f_p)[0]
-
-    return "%s_fastqc.html" % f_p, "%s_fastqc.zip" % f_p
-
-
-
-def _run_commands(qclient, job_id, commands, msg):
-    for i, cmd in enumerate(commands):
-        qclient.update_job_step(job_id, msg % i)
-        std_out, std_err, return_value = system_call(cmd)
-        if return_value != 0:
-            error_msg = ("Error running FastQC:\nStd out: %s\nStd err: %s"
-                         "\n\nCommand run was:\n%s"
-                         % (std_out, std_err, cmd))
-            return False, error_msg
-
-    return True, ""
-
-
-def fastqc(qclient, job_id, parameters, out_dir):
-    """Run FastQC and MultiQC on a sequence artifact
-
-    Parameters
-    ----------
-    qclient : tgp.qiita_client.QiitaClient
-        The Qiita server client
-    job_id : str
-        The job id
-    parameters : dict
-        The parameter values to run split libraries
-    out_dir : str
-        Yhe path to the job's output directory
-
-    Returns
-    -------
-    bool, list, str
-        The results of the job
-    """
-    # Step 1 get the rest of the information need to run FastQC
-    qclient.update_job_step(job_id, "Step 1 of 3: Collecting information")
-    artifact_id = parameters['input']
-
-    # removing input from parameters so it's not part of the final command
-    del parameters['input']
-
-    # Get the artifact filepath information
-    artifact_info = qclient.get("/qiita_db/artifacts/%s/" % artifact_id)
-    fps = artifact_info['files']
-
-    # Get the artifact metadata
-    prep_info = qclient.get('/qiita_db/prep_template/%s/'
-                            % artifact_info['prep_information'][0])
-    qiime_map = prep_info['qiime-map']
-
-    # Generate temp dir for exection of QC steps
-    out_dir = mkdtemp()
-
-    # Step 2 generating command FastQC
-    qclient.update_job_step(job_id, "Step 2 of 3: Generating"
-                                    " FastQC command")
-
-    fqc_out_dir = join(out_dir, 'FastQC')
-    
-    rs = fps['raw_reverse_seqs'] if 'raw_reverse_seqs' in fps else []
-    fqc_cmds, samples = generate_fastqc_commands(fps['raw_forward_seqs'],
-                                                 rs, qiime_map, fqc_out_dir,
-                                                 parameters)
-
-    # Step 3 execute FastQC
-    msg = "Step 3 of 3: Executing FastQC job (%d/{0})".format(len(fqc_cmds))
-    success, msg = _run_commands(qclient, job_id, fqc_cmds, msg)
-    if not success:
-        return False, None, msg
-
-    # Generate MultiQC command
-    mqc_out_dir = join(out_dir, 'MultiQC')
-
-    qclient.update_job_step(job_id, "Step 4 of 3: Generating"
-                                    " MultiQC command")
-    
-    mqc_cmds, samples = generate_multiqc_commands(mqc_out_dir,
-                                                  parameters)
-
-    # Execute MultiQC
-    msg = "Step 5 of 3: Executing MultiQC job (%d/{0})".format(len(mqc_cmds))
-    success, msg = _run_commands(qclient, job_id, mqc_cmds, msg)
-    if not success:
-        return False, None, msg
-
-    # Compress QC files
-    commands = []
-
-    fqc_cmd = "tar -czvf %s %s" % (join(out_dir,'fastqc.tar.gz'), fqc_out_dir)
-    mqc_cmd = "tar -czvf %s %s" % (join(out_dir,'multiqc.tar.gz'), mqc_out_dir)
-
-    msg = "Step 6 of 3: Compressing FastQC and MultiQC output"
-    success, msg = _run_commands(qclient, job_id, commands, msg)
-    if not success:
-        return False, None, msg
-    
-
-    # Step 4 generating artifacts
-
-    ainfo = []
-
-    ainfo.extend([
-        ArtifactInfo('MultiQC data', 'tgz',
-                     [(join(out_dir,'multiqc.tar.gz'), 'tgz')]),
-        ArtifactInfo('FastQC data', 'tgz',
-                     [(join(out_dir,'fastqc.tar.gz'), 'tgz')])])
-
-
-    return True, ainfo, ""
